Refactor login page route, try to write a spec for it
