+++ src/apps/loyalty/__mocks__/artsy-passport.ts
@@ -1,0 +1,6 @@
+module.exports = {
+  options: {
+    facebookPath: "/facebook",
+    twitterPath: "/twitter",
+  },
+}
