+++ src/apps/loyalty/__tests__/server.tsx
@@ -1,4 +1,4 @@
-import { Home } from "../server/route_handlers"
+import { Home, Login } from "../server/route_handlers"
 
 describe ("Home page", () => {
   it("redirects to /inquiries", () => {
@@ -7,5 +7,15 @@ describe ("Home page", () => {
     let next = jest.fn() as any
     Home(req, res, next)
     expect(res.redirect).toHaveBeenCalledWith("loyalty/inquiries")
+  })
+})
+
+describe ("Login page", () => {
+  it("renders the login page", () => {
+    let req = { baseUrl: "loyalty", csrfToken: () => "token" } as any
+    let res = { redirect: jest.fn(), send: jest.fn() } as any
+    let next = jest.fn() as any
+    Login(req, res, next)
+    expect(res.send).toHaveBeenCalledWith("test")
   })
 })
